+++ src/clldutils/__init__.py
@@ -1,1 +1,4 @@
+"""
+Utilities useful for managing/publishing research data.
+"""
 __version__ = '3.24.5.dev0'