+++ src/clldutils/misc.py
@@ -6,25 +6,27 @@ functionality (:func:`lazyproperty`), formatting functions, etc.
 import re
 import base64
 import string
-import typing
+from typing import Union, Any
 import pathlib
 import warnings
 import mimetypes
 import unicodedata
+from collections.abc import Iterable
 
 __all__ = [
     'data_url', 'log_or_raise', 'nfilter', 'to_binary', 'dict_merged', 'NoDefault', 'NO_DEFAULT',
-    'xmlchars', 'format_size', 'UnicodeMixin', 'slug', 'encoded', 'lazyproperty',
+    'xmlchars', 'format_size', 'slug', 'encoded',
 ]
 
 
-def deprecated(msg):
+def deprecated(msg):  # pragma: no cover
+    """Mark deprecated functionality."""
     warnings.simplefilter('always', DeprecationWarning)
     warnings.warn(msg, category=DeprecationWarning, stacklevel=2)
     warnings.simplefilter('default', DeprecationWarning)
 
 
-def data_url(content: typing.Union[bytes, str, pathlib.Path], mimetype: str = None) -> str:
+def data_url(content: Union[bytes, str, pathlib.Path], mimetype: str = None) -> str:
     """
     Returns content encoded as base64 Data URI. Useful to include (smallish) media resources
     in HTML pages.
@@ -43,8 +45,8 @@ def data_url(content: typing.Union[bytes, str, pathlib.Path], mimetype: str = No
     else:
         if isinstance(content, str):
             content = content.encode('utf8')
-    return "data:{0};base64,{1}".format(
-        mimetype or 'application/octet-stream', base64.b64encode(content).decode())
+    mimetype = mimetype or 'application/octet-stream'
+    return f"data:{mimetype};base64,{base64.b64encode(content).decode()}"
 
 
 def log_or_raise(msg: str, log=None, level='warning', exception_cls=ValueError):
@@ -69,7 +71,7 @@ def log_or_raise(msg: str, log=None, level='warning', exception_cls=ValueError):
         raise exception_cls(msg)
 
 
-def nfilter(seq: typing.Iterable) -> list:
+def nfilter(seq: Iterable[Any]) -> list[Any]:
     """Replacement for python 2's filter(None, seq).
 
     :return: a list filtered from seq containing only truthy items.
@@ -77,7 +79,7 @@ def nfilter(seq: typing.Iterable) -> list:
     return [e for e in seq if e]
 
 
-def to_binary(s: typing.Union[str, bytes], encoding='utf8') -> bytes:
+def to_binary(s: Union[str, bytes], encoding='utf8') -> bytes:
     """Cast function.
 
     :param s: object to be converted to bytes.
@@ -106,8 +108,8 @@ def dict_merged(d, _filter=None, **kw):
     return d
 
 
-class NoDefault(object):
-
+class NoDefault:  # pylint: disable=too-few-public-methods
+    """A default object for cases, where `None` is considered a regular value."""
     def __repr__(self):
         return '<NoDefault>'
 
@@ -127,7 +129,8 @@ def xmlchars(text: str) -> str:
     invalid = list(range(0x9))
     invalid.extend([0xb, 0xc])
     invalid.extend(range(0xe, 0x20))
-    return re.sub('|'.join('\\x%0.2X' % i for i in invalid), '', text)
+    return re.sub(
+        '|'.join('\\x%0.2X' % i for i in invalid), '', text)  # pylint: disable=C0209
 
 
 def format_size(num: int) -> str:
@@ -141,24 +144,11 @@ def format_size(num: int) -> str:
 
     .. seealso:: `<http://stackoverflow.com/a/1094933>`_
     """
-    for x in ['bytes', 'KB', 'MB', 'GB']:
-        if num < 1024.0 and num > -1024.0:
-            return "%3.1f%s" % (num, x)
+    for x in ['bytes', 'KB', 'MB', 'GB', 'TB']:
+        if -1024.0 < num < 1024.0:
+            return f"{num:3.1f}{x}"
         num /= 1024.0
-    return "%3.1f%s" % (num, 'TB')
-
-
-class UnicodeMixin(object):
-    """Portable label mixin."""
-
-    def __unicode__(self):
-        """a human readable label for the object."""
-        return '%s' % self  # pragma: no cover
-
-    def __str__(self):
-        """a human readable label for the object, appropriately encoded (or not)."""
-        deprecated("Use of deprecated class UnicodeMixin! Use object instead.")
-        return self.__unicode__()
+    return f"{num:3.1f}PB"
 
 
 def slug(s: str, remove_whitespace: bool = True, lowercase: bool = True) -> str:
@@ -187,56 +177,19 @@ def slug(s: str, remove_whitespace: bool = True, lowercase: bool = True) -> str:
     return res
 
 
-def encoded(string: typing.Union[str, bytes], encoding='utf-8') -> bytes:
+def encoded(string_: Union[str, bytes], encoding='utf-8') -> bytes:
     """Cast string to bytes in a specific encoding - with some guessing about the encoding.
 
     :param encoding: encoding which the object is forced to
     """
-    assert isinstance(string, (str, bytes))
-    if isinstance(string, str):
-        return string.encode(encoding)
+    assert isinstance(string_, (str, bytes))
+    if isinstance(string_, str):
+        return string_.encode(encoding)
     try:
         # make sure the string can be decoded in the specified encoding ...
-        string.decode(encoding)
-        return string
+        string_.decode(encoding)
+        return string_
     except UnicodeDecodeError:
         # ... if not use latin1 as best guess to decode the string before encoding as
         # specified.
-        return string.decode('latin1').encode(encoding)
-
-
-class lazyproperty(object):
-    """Non-data descriptor caching the computed result as instance attribute.
-
-    .. code-block:: python
-
-        >>> class Spam(object):
-        ...     @lazyproperty
-        ...     def eggs(self):
-        ...         return 'spamspamspam'
-        >>> spam=Spam(); spam.eggs
-        'spamspamspam'
-        >>> spam.eggs='eggseggseggs'; spam.eggs
-        'eggseggseggs'
-        >>> Spam().eggs
-        'spamspamspam'
-        >>> Spam.eggs  # doctest: +ELLIPSIS
-        <...lazyproperty object at 0x...>
-
-    .. note::
-
-        Since Python 3.8 added the `functools.cached_property` decorator
-        (see `<https://docs.python.org/3/library/functools.html#functools.cached_property>`_),
-        this function will be deprecated once Python 3.7 is no longer supported.
-    """
-
-    def __init__(self, fget):
-        self.fget = fget
-        for attr in ('__module__', '__name__', '__doc__'):
-            setattr(self, attr, getattr(fget, attr))
-
-    def __get__(self, instance, owner):
-        if instance is None:
-            return self
-        result = instance.__dict__[self.__name__] = self.fget(instance)
-        return result
+        return string_.decode('latin1').encode(encoding)