+++ tests/test_clilib.py
@@ -80,84 +80,6 @@ def test_register_subcommands_error(fixtures_dir, mocker, recwarn):
         assert recwarn.pop(UserWarning)
 
 
-def test_ArgumentParser(capsys):
-    def cmd(args):
-        """
-        docstring
-        """
-        if len(args.args) < 1:
-            raise ParserError('not enough arguments')
-        print(args.args[0])
-
-    parser = ArgumentParserWithLogging('pkg', cmd)
-
-    parser.main(args=['help', 'cmd'])
-    out, err = capsys.readouterr()
-    assert 'docstring' in out
-
-    parser.main(args=['cmd', 'arg'])
-    out, err = capsys.readouterr()
-    assert 'arg' in out
-
-    assert parser.main(args=['cmd', 'arg']) == 0
-
-    parser.main(args=['cmd'])
-    out, err = capsys.readouterr()
-    assert 'not enough arguments' in out
-
-    assert parser.main(args=['x']) != 0
-    out, err = capsys.readouterr()
-    assert out.startswith('invalid')
-
-    @command()
-    def ls(args):
-        """
-        my name is ls
-        """
-        return
-
-    @command(name='list', usage='my name is {0}'.format('list'))
-    def f(args):
-        """
-        """
-        return
-
-    parser = ArgumentParserWithLogging('pkg')
-    parser.main(args=['help', 'ls'])
-    out, err = capsys.readouterr()
-    assert 'my name is ls' in out
-
-    parser.main(args=['help', 'list'])
-    out, err = capsys.readouterr()
-    assert 'my name is list' in out
-
-    assert parser.main(args=['ls', 'arg']) == 0
-    assert parser.main(args=['list', 'arg']) == 0
-
-
-def test_deprecation(recwarn):
-    warnings.simplefilter("always")
-
-    class AP(ArgumentParser):
-        pass
-
-    assert recwarn.pop(DeprecationWarning)
-    warnings.simplefilter("default")
-
-
-def test_cmd_error():
-    from clldutils.clilib import ArgumentParser
-
-    def cmd(args):
-        raise ValueError
-
-    parser = ArgumentParser('pkg', cmd)
-    with pytest.raises(ValueError):
-        parser.main(args=['cmd'])
-
-    assert parser.main(args=['cmd'], catch_all=True) == 1
-
-
 def test_confirm(capsys, mocker):
     from clldutils.clilib import confirm
 
@@ -175,7 +97,7 @@ def test_Table(capsys):
     with Table(argparse.Namespace(format='simple'), 'a') as t:
         t.append(['x'])
     out, _ = capsys.readouterr()
-    assert out == 'a\n---\nx\n'
+    assert out == ' a  \n---\n x  \n'
 
 
 def test_add_format():