+++ tests/test_markup.py
@@ -8,19 +8,29 @@ from markdown import markdown
 from clldutils.markup import *
 
 
-def test_Table():
+@pytest.mark.parametrize(
+    'options,expected',
+    [
+        (dict(),
+         '| a | b |\n| :-: | :-: |\n| 1 | 2 |\n| 3 | 4 |'),
+        (dict(condensed=False),
+         '|  a  |  b  |\n| :-: | :-: |\n|  1  |  2  |\n|  3  |  4  |'),
+        (dict(verbose=True),
+         '| a | b |\n| :-: | :-: |\n| 1 | 2 |\n| 3 | 4 |\n\n(2 rows)\n\n'),
+        (dict(sortkey=itemgetter(1), reverse=True),
+         '| a | b |\n| :-: | :-: |\n| 3 | 4 |\n| 1 | 2 |'),
+    ]
+)
+def test_Table(options, expected):
     t = Table()
     assert t.render() == ''
 
     t = Table('a', 'b', rows=[[1, 2], [3, 4]])
-    assert t.render() == \
-        '| a | b |\n|----:|----:|\n| 1 | 2 |\n| 3 | 4 |'
-    assert t.render(condensed=False) == \
-        '|   a |   b |\n|----:|----:|\n|   1 |   2 |\n|   3 |   4 |'
-    assert t.render(verbose=True) == \
-        '| a | b |\n|----:|----:|\n| 1 | 2 |\n| 3 | 4 |\n\n(2 rows)\n\n'
-    assert t.render(sortkey=itemgetter(1), reverse=True) == \
-        '| a | b |\n|----:|----:|\n| 3 | 4 |\n| 1 | 2 |'
+    assert t.render(**options) == expected
+
+
+def test_TableFormat():
+    assert TableFormat.get(TableFormat.pipe) == TableFormat.pipe
 
 
 def test_Table_tsv(capsys):
@@ -32,19 +42,24 @@ def test_Table_tsv(capsys):
 
 def test_Table_context(capsys):
     with Table('a', 'b', tablefmt='simple') as t:
+        pass
+    out, _ = capsys.readouterr()
+    assert ' a ' in out
+
+    with Table('a', 'b', 'b', tablefmt='simple') as t:
         t.append([1, 2.345])
     out, _ = capsys.readouterr()
-    assert out == '  a     b\n---  ----\n  1  2.35\n'
-
+    assert out == ' a   b      b_2  \n--- ------ -----\n 1   2.35        \n'
     f = io.StringIO()
     with Table('a', 'b', tablefmt='simple', file=f) as t:
         t.append([1, 2.345])
-    assert f.getvalue() == '  a     b\n---  ----\n  1  2.35\n'
+    assert f.getvalue() == ' a   b     \n--- ------\n 1   2.35  \n'
 
 
 def test_iter_markdown_tables():
     header, rows = ['a', 'b'], [[1, 2], [3, 4]]
     text = Table(*header, **dict(rows=rows)).render() + '\nabcd'
+    print(text)
     assert list(iter_markdown_tables(text))[0] == \
         (header, [[str(v) for v in r] for r in rows])
     assert list(iter_markdown_tables('a|b\n---|---\n1|2'))[0] == (header, [['1', '2']])
@@ -99,6 +114,17 @@ def test_add_markdown_text_error():
 )
 def test_markdownlink1(markdown, expected):
     assert MarkdownLink.replace(markdown, lambda ml: ml.update_url(fragment='c')) == expected
+
+def test_non_link_square_brackets():
+    assert MarkdownLink.replace(
+        '\\[stuff\\] and [a](b)) ', lambda ml: "--") == '\\[stuff\\] and --) '
+    assert MarkdownLink.replace(
+        '[stuff] and [a](b)) ', lambda ml: "--") == '[stuff] and --) '
+
+
+def test_invalid_markdownlink():
+    with pytest.raises(ValueError):
+        _ = MarkdownLink.from_string('[abc]')
 
 
 def test_non_link_square_brackets():